se_backend/requirements.txt
alabaster==0.7.12
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
Babel==2.8.0
beautifulsoup4==4.9.3
black==19.10b0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
comtypes==1.1.7
csvvalidator==1.2
cx-Freeze==6.2
Django==3.0.8
django-audit-log==0.7.0
django-auditlog==0.4.7
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.4.0
django-crum==0.7.7
django-currentuser==0.5.1
django-excel==0.0.10
django-filter==2.3.0
django-jsonfield==1.4.1
django-oauth-toolkit==1.3.2
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
docutils==0.16
drf-renderer-xlsx==0.3.7
et-xmlfile==1.0.1
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.10
imagesize==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
joblib==0.16.0
lml==0.0.9
MarkupSafe==1.1.1
nltk==3.5
numpy==1.19.1
oauthlib==3.1.0
openpyxl==3.0.4
packaging==20.4
pandas==1.0.5
pathspec==0.8.0
Pillow==7.2.0
psycopg2==2.8.5
pyexcel==0.6.2
pyexcel-io==0.5.20
pyexcel-webio==0.1.4
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pypiwin32==223
python-dateutil==2.6.0
pyttsx3==2.90
pytz==2020.1
pywin32==228
regex==2020.7.14
requests==2.24.0
sentry-sdk==0.18.0
six==1.15.0
snowballstemmer==2.0.0
soupsieve==2.0.1
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.3.18
sqlparse==0.3.1
texttable==1.6.2
toml==0.10.1
tqdm==4.48.0
twilio==6.44.1
typed-ast==1.4.1
urllib3==1.25.10
xlrd==1.2.0
XlsxWriter==1.2.9
